add example of fhe (base64) for json de-crpt.
I added the serde_json library for JSON operations.
I added the base64 library, we will use it as a simple encoding to represent FHE encryption.
I added a new function called fhe_encrypt. This function uses a simple Base64 encoding rather than a true FHE implementation. For a real FHE implementation, you will need to use a special library.
In the main function, we generate the results in JSON format and then encrypt this JSON string with FHE (Base64 encoding is used in this example).
We generate and encrypt JSON output for both successful results and error cases.